refactor(demo): Behaviors directory Migrate add-styles directory in its own subdir, to match the structure adopted for Components directory.
