+++ demo/src/Behaviors/AddStyles/index.ts
@@ -1,0 +1,1 @@
+export { AddStyles } from './AddStyles';
